project-bolt-sb1-tzqh524d/project/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.6.0
python-multipart==0.0.6
httpx==0.26.0
